added bbs link + mp4
Saw the .mp4 in the bbs thread. The .mp4 is really useful, because you can pause it, add a bookmark and resume later, etc.

You might want to mirror it, or upload it on Youtube. And thanks for that tutorial, currently doing it, and really enjoying your presentation and ideas.